Add utility method for generating AnnotatedType without annotation
