+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/property/TupleLikeElementsProperty.java
@@ -18,6 +18,8 @@
 
 package com.navercorp.fixturemonkey.api.property;
 
+import static com.navercorp.fixturemonkey.api.type.Types.generateAnnotatedTypeWithoutAnnotation;
+
 import java.lang.annotation.Annotation;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Type;
@@ -41,28 +43,7 @@ public final class TupleLikeElementsProperty implements Property {
 
 	private final Type type = TupleLikeElementsType.class;
 
-	private final AnnotatedType annotatedType = new AnnotatedType() {
-		@Override
-		public Type getType() {
-			return type;
-		}
-
-		@Nullable
-		@Override
-		public <T extends Annotation> T getAnnotation(Class<T> annotationClass) {
-			return null;
-		}
-
-		@Override
-		public Annotation[] getAnnotations() {
-			return new Annotation[0];
-		}
-
-		@Override
-		public Annotation[] getDeclaredAnnotations() {
-			return new Annotation[0];
-		}
-	};
+	private final AnnotatedType annotatedType = generateAnnotatedTypeWithoutAnnotation(type);
 
 	public TupleLikeElementsProperty(
 		Property tupleLikeProperty,
